move more artefacts to java7 project

+++ troilus-core-minimal/src/main/java/com/unitedinternet/troilus/MinimalDeleteQuery.java
@@ -41,7 +41,7 @@ class MinimalDeleteQuery extends AbstractQuery<Deletion> implements Deletion {
     }
 
 
-    public Deletion withTtl(long ttlSec) {
+    public Deletion withTtl(int ttlSec) {
         return newQuery(getContext().withTtl(ttlSec));
     }
